pyproject.toml
[project]
name = "webmonstro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "black>=25.1.0",
    "colorama>=0.4.6",
    "fastapi>=0.115.12",
    "matplotlib>=3.10.1",
    "orjson>=3.10.16",
    "pandas>=2.2.3",
    "pydantic-settings>=2.8.1",
    "pyyaml>=6.0.2",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]
